requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
lxml==4.9.3
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.7.1
urllib3==2.0.7
yarl==1.9.2
APScheduler~=3.10.4